feat: add a semantic name for internal exception for a better dev experience on unwrap_or_return + @early_return feature
